requirements.txt
fastapi
hypercorn
tensorflow
numpy
pillow
gdown
keras